Implement automatic history retrieval for unread IMs upon login.
It will automatically open conversation windows for all direct and
multi-party direct messages (IM and MPIM), but not for other
channels. It is assumed that the user will want to join other channels
separately.

There is no documented API for getting unread status of all chats
simultaneously, but for efficiency reasons this is important to avoid
getting throttled. So instead we use the undocumented "users.counts"
API call to query for this. Thanks to EionRobb (GitHub) for finding
this one!

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>