Switch unread message retrieval to queue-based/rate-limited approach.
Use a queue to hold unread message IDs we want to retrieve and trigger
an update every five seconds. This is important to avoid black listing
from Slack if there is a large number of messages, and it's pretty
unlikely that the user will be able to read and process messages
faster than five seconds each.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>