Add ability to fetch thread history.
To do this we have to get around some serious deficiences in the
public API: There no way to query replies by date. And we have to do
that in order to get automatic history when you log in. The only way I
have found is to query as many channel messages as possible, check
what their "latest_reply" field is, and use those to decide what
threads to query. Checking all messages this way is completely
unrealistic, but fortunately, threads tend to be short-lived, so we
can use one single API call to get the last 1000 messages (maximum
that Slack allows), and check that without much risk of being rate
limited. This means the mechanics for querying by time is (mostly)
moved to the client side, and when limiting by time, we still fetch
all the last 1000 messages, but we only display those that match the
time range.

Yes, this is horrible, and it will not work for any thread older than
1000 main-channel messages ago. If anyone knows how to query threads
in a more efficient way, I'm very interested in hearing it.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>