Add ability to post to Slack threads.
The way it's implemented in Purple is more like topics. The new chat
command "/thread", or "/th" for short, allows you to post to threads
by using their timestamp as an identifier, like how you would mention
someone's name or a group. You can also switch in and out of thread
context, causing all subsequent messages to go to that thread, which
is handy for busy thread discussions.

Posting to threads works in both channels and IMs, but in IMs it
behaves a bit strangely, and this is mostly due to how Purple works
(at least when using Pidgin). When switching to a thread context, and
then posting, you will see both a thread-less and then a threaded
message. This is because when you are not using a command, Purple
posts the message immediately, without letting the plugin control how
it should be displayed. Only once the server replies does it get
displayed properly. I'm not sure if this is fixable, but since threads
are less common in IMs, I'm not rushing to fix it.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>