Fix truncated messages when editing long messages.
The problem manifests when you are trying to edit a long message. The
partial URL mode causes the request, and therefore the message, to be
truncated. A message with a 100 characters is more than enough to
trigger it.

Therefore switch to POST so we can include the long parameters in the
body instead. We take the opportunity to switch all appropriate
endpoints over to POST, which is a more correct way to use the API in
general.

It's a bit hairy to do POST requests since Purple does not support
them natively, but we just need to construct a small set of headers.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>