Improve timestamp display: Less noise and more options.
Don't display the empty diamonds for channel messages that don't have
any thread replies. This is just too verbose and detracts from the
color coding. Instead, provide an option to redisplay messages when a
thread is opened on them, so that the color and context is clearly
visible.

If the option is off, channel messages are completely unaffected by
thread replies. If the thread replies option is also off, then there
will be no indication of thread replies whatsoever.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>