Fix unread history bug when threads are mixed with channel messages.
The problem only happens when either fetching manually or when opening
a channel that has not fetched unread history as part of logging in.
It does not affect automatic unread history fetching. It happens
because we are only fetching the number of messages which is in the
unread count, but this includes threaded replies. Since threaded
replies are not included in normal history queries, we need to do like
we do when fetching history automatically: Fetch as much history as we
can, and then match timestamps locally to figure out which messages to
display, and which threads to query.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>