Display parent message when /getthread is used
This commit displays the parent message of the thread when the `/getthread`
command is used, because it often has context in the parent message that
the thread participants assume you can already see. I almost always want
to see the parent message when I run `getthread`, and when I can't see
it, I find my self running `/history 10`, then `/history 100`, etc.
until I can see the parent message and match it back to the thread
replies, which is annoying. This fixed it for me, and it hopefully also
useful for others :)