slack-message: add option for dropping old messages
Sometimes a user (or bot) edits or deletes a message that is quite
old. It can be very confusing to see such messages inserted in the
current timeline, even with the '[edit]' or 'Deleted message' prefix,
because they are totally out of context.

This adds an account option that allows hiding of edits/deletes on
messages that are older than a certain number of hours. The option
defaults to 0, so users get the full stream unless opting out.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>